chore: Refactor namespaces.
