+++ WebEditor/Components/Board/BoardRenderer.razor
@@ -1,4 +1,3 @@
-@using CaptainCoder.TacticsEngine.Board
 @using System.Diagnostics.CodeAnalysis;
 @inject ToolManager ToolManager;
 <div class="board-container" style="--cell-size: @(CellSize)px;
