+++ WebEditor/Components/Board/FigureRenderer.razor
@@ -1,4 +1,3 @@
-@using CaptainCoder.TacticsEngine.Board
 @using System.Diagnostics.CodeAnalysis
 @inject ToolManager ToolManager
