+++ WebEditor/Components/Board/PositionedFigureRenderer.razor
@@ -1,6 +1,4 @@
-@using CaptainCoder.TacticsEngine.Board
 @using System.Diagnostics.CodeAnalysis
-@using WebEditor.Components.BoardEditor.Panel
 @inject ToolManager ToolManager
 @if (IsVisible)
 {
