+++ WebEditor/Components/Board/PositionedFigureRenderer.razor.css
