+++ WebEditor/Components/Board/TileRenderer.razor
@@ -1,5 +1,3 @@
-@using CaptainCoder.TacticsEngine.Board
-@using System.Diagnostics.CodeAnalysis
 @inject ToolManager ToolManager
 <div 
     class="tile-info @TileType"
