+++ WebEditor/Components/Editors/BoardEditor.razor
@@ -1,9 +1,7 @@
 @using System.Text.Json
 @using Blazor.DownloadFileFast.Interfaces
-@using CaptainCoder.TacticsEngine.Board
 @using System.Diagnostics.CodeAnalysis
 @using System.Text
-@using WebEditor.Components.BoardEditor.Panel
 @inject IJSRuntime JSRuntime
 @inject IBlazorDownloadFileService BlazorDownloadFileService
 @inject ToolManager ToolManager
