+++ WebEditor/Components/Panels/FigurePanel.razor
