+++ WebEditor/Components/Panels/FigurePanel.razor.css
