+++ WebEditor/Components/Tools/TileTool.cs
@@ -1,6 +1,6 @@
 using CaptainCoder.TacticsEngine.Board;
 
-namespace WebEditor.Components.BoardEditor;
+namespace WebEditor.Tools;
 
 public sealed class TileTool : Tool
 {
