+++ WebEditor/Pages/Home.razor
@@ -1,5 +1,4 @@
 ﻿@page "/"
-@using WebEditor.Components.BoardEditor;
 
 <PageTitle>Tactics Board Editor</PageTitle>
