+++ WebEditor/Program.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Components.Web;
 using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
 
-using WebEditor.Components.BoardEditor;
+using WebEditor.Tools;
 
 var builder = WebAssemblyHostBuilder.CreateDefault(args);
 builder.RootComponents.Add<WebEditor.App>("#app");
