+++ WebEditor/_Imports.razor
@@ -8,3 +8,10 @@
 @using Microsoft.JSInterop
 @using WebEditor
 @using WebEditor.Layout
+@using WebEditor.Components
+@using WebEditor.Components.Board
+@using WebEditor.Components.Editors
+@using WebEditor.Components.Panels
+@using WebEditor.Tools
+@using CaptainCoder.TacticsEngine.Board
+@using System.Diagnostics.CodeAnalysis
